has finished the communication of GWs, but something wrong with the data

+++ modules/messages/WaveShortMessage_m.h
@@ -39,7 +39,7 @@ struct Info
 };
 
 struct Connectivity_Info {
-	double connectivityValue;
+	int connectivityValue;
 	LAddress::L2Type macAddrGW;
 };
